Moving FromJSONObject functions to 3party/jansson/myjansson unit.
