+++ 3party/jansson/jansson.pro
@@ -19,6 +19,7 @@ SOURCES += \
     src/error.c \
     src/strconv.c \
     jansson_handle.cpp \
+    myjansson.cpp \
 
 HEADERS += \
     myjansson.hpp \
