+++ 3party/jansson/myjansson.hpp
@@ -3,8 +3,11 @@
 #include "jansson_handle.hpp"
 
 #include "base/exception.hpp"
+#include "base/string_utils.hpp"
 
 #include <jansson.h>
+
+#include "std/vector.hpp"
 
 namespace my
 {
@@ -25,4 +28,23 @@ public:
 
   json_t * get() const { return m_handle.get(); }
 };
+
+void FromJSON(json_t * root, string & result);
+void FromJSONObject(json_t * root, string const & field, string & result);
+void FromJSONObject(json_t * root, string const & field, strings::UniString & result);
+void FromJSONObject(json_t * root, string const & field, double & result);
+
+template <typename T>
+void FromJSONObject(json_t * root, string const & field, vector<T> & result)
+{
+  json_t * arr = json_object_get(root, field.c_str());
+  if (!arr)
+    MYTHROW(my::Json::Exception, ("Obligatory field", field, "is absent."));
+  if (!json_is_array(arr))
+    MYTHROW(my::Json::Exception, ("The field", field, "must contain a json array."));
+  size_t sz = json_array_size(arr);
+  result.resize(sz);
+  for (size_t i = 0; i < sz; ++i)
+    FromJSON(json_array_get(arr, i), result[i]);
+}
 }  // namespace my
